4088: Fixed circular dependencies for native
