+++ native/src/components/StaticServerProvider.tsx
@@ -1,7 +1,7 @@
 import StaticServer from '@dr.pogodin/react-native-static-server'
 import React, { createContext, ReactElement, ReactNode, useEffect, useState } from 'react'
 
-import { RESOURCE_CACHE_DIR_PATH } from '../utils/DatabaseConnector'
+import { RESOURCE_CACHE_DIR_PATH } from '../constants/DatabasePaths'
 import { reportError } from '../utils/sentry'
 
 type StaticServerProps = {
