+++ native/src/utils/DatabaseConnector.ts
@@ -22,22 +22,17 @@ import {
   OfferModel,
 } from 'shared/api'
 
+import {
+  CACHE_DIR_PATH,
+  CONTENT_DIR_PATH,
+  RESOURCE_CACHE_DIR_PATH,
+  UNVERSIONED_CONTENT_DIR_PATH,
+  UNVERSIONED_RESOURCE_CACHE_DIR_PATH,
+} from '../constants/DatabasePaths'
 import DatabaseContext from '../models/DatabaseContext'
 import { RegionResourceCacheStateType } from './DataContainer'
 import { deleteIfExists } from './helpers'
 import { log, reportError } from './sentry'
-
-export const CONTENT_VERSION = 'v12'
-export const RESOURCE_CACHE_VERSION = 'v4'
-
-// Our pdf view can only load from DocumentDir. Therefore we need to use that
-export const CACHE_DIR_PATH = BlobUtil.fs.dirs.DocumentDir
-export const UNVERSIONED_CONTENT_DIR_PATH = `${CACHE_DIR_PATH}/content`
-// Offline saved content like categories, events and pois
-export const CONTENT_DIR_PATH = `${UNVERSIONED_CONTENT_DIR_PATH}/${CONTENT_VERSION}`
-export const UNVERSIONED_RESOURCE_CACHE_DIR_PATH = `${CACHE_DIR_PATH}/resource-cache`
-// Offline saved resources like pictures and pdf documents
-export const RESOURCE_CACHE_DIR_PATH = `${UNVERSIONED_RESOURCE_CACHE_DIR_PATH}/${RESOURCE_CACHE_VERSION}`
 
 const MAX_STORED_REGIONS = 3
