+++ native/src/utils/createSettingsSections.ts
@@ -12,7 +12,8 @@ import { NavigationProps } from '../constants/NavigationTypes'
 import buildConfig from '../constants/buildConfig'
 import { AppContextType } from '../contexts/AppContextProvider'
 import { SettingsType } from './AppSettings'
-import { requestPushNotificationPermission, subscribeNews, unsubscribeNews } from './PushNotificationsManager'
+import { requestPushNotificationPermission } from './PushNotificationsManager'
+import { subscribeNews, unsubscribeNews } from './PushNotificationsNews'
 import openExternalUrl from './openExternalUrl'
 import { initSentry } from './sentry'
